Break up payment-flow.tsx into co-located step files
Split the 557-line monolith into focused components under a
_components/ private folder, following the Next.js App Router
convention for route-private files.
