+++ apps/web/lib/near-intents/index.ts
@@ -1,4 +1,5 @@
 export { getSwapQuote } from './quote';
 export type { QuoteResponse } from './quote';
-export { getSwapStatus, GetExecutionStatusResponse } from './status';
+export { getSwapStatus } from './status';
+export type { GetExecutionStatusResponse } from './status';
 export { getDefuseAssetId, toSmallestUnits, getTokenDecimals } from './assets';
