+++ apps/web/package.json
@@ -14,6 +14,7 @@
     "@defuse-protocol/one-click-sdk-typescript": "^0.1.15",
     "@repo/database": "workspace:*",
     "@t3-oss/env-nextjs": "^0.13.10",
+    "@tanstack/react-query": "^5.90.21",
     "better-auth": "^1.2.8",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
